Add function to run emulator from command line
This allows me to open and run an AVD from the command line. However I
kept running into an issue because the emulator command was trying to
use the emulator.exe inside the tools folder, which is not working. So
a solution from here:  https://stackoverflow.com/a/44822312/8104787
suggested moving $ANDROID_SDK_HOME\emulator in the line before the
tools directory, this way the emulator executable will be searched
inside the emulator folder first and will take precedence over the one
present in the tools folder