Add a Gif Function To Convert Android Screenrecordings directly to GIFS
This uses gifify(https://github.com/vvo/gifify). I have it set up to grab the recording off the device, move it to the desktop and then convert it to a GIF. Ideally we could convert the mp4 directly in the remote shell, but installing gifify in the remote shell is less than ideal. If that's not neccessary, and I learn how to improve this one day, that will be a great day.